Move Phase class to processing package
