+++ src/main/java/fi/digitraffic/tis/vaco/process/PhaseService.java
@@ -1,7 +1,7 @@
 package fi.digitraffic.tis.vaco.process;
 
 import fi.digitraffic.tis.utilities.model.ProcessingState;
-import fi.digitraffic.tis.vaco.queuehandler.model.ImmutablePhase;
+import fi.digitraffic.tis.vaco.process.model.ImmutablePhase;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
