+++ src/main/java/fi/digitraffic/tis/vaco/process/model/PhaseData.java
@@ -1,6 +1,5 @@
 package fi.digitraffic.tis.vaco.process.model;
 
-import fi.digitraffic.tis.vaco.queuehandler.model.ImmutablePhase;
 import jakarta.annotation.Nullable;
 import org.immutables.value.Value;
