+++ src/main/java/fi/digitraffic/tis/vaco/queuehandler/model/Entry.java
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import fi.digitraffic.tis.vaco.DataVisibility;
 import fi.digitraffic.tis.vaco.errorhandling.Error;
+import fi.digitraffic.tis.vaco.process.model.Phase;
 import jakarta.annotation.Nullable;
 import org.immutables.value.Value;
