+++ src/main/java/fi/digitraffic/tis/vaco/queuehandler/repository/QueueHandlerRepository.java
@@ -9,9 +9,9 @@ import fi.digitraffic.tis.vaco.queuehandler.model.ConversionInput;
 import fi.digitraffic.tis.vaco.queuehandler.model.Entry;
 import fi.digitraffic.tis.vaco.queuehandler.model.ImmutableConversionInput;
 import fi.digitraffic.tis.vaco.queuehandler.model.ImmutableEntry;
-import fi.digitraffic.tis.vaco.queuehandler.model.ImmutablePhase;
+import fi.digitraffic.tis.vaco.process.model.ImmutablePhase;
 import fi.digitraffic.tis.vaco.queuehandler.model.ImmutableValidationInput;
-import fi.digitraffic.tis.vaco.queuehandler.model.Phase;
+import fi.digitraffic.tis.vaco.process.model.Phase;
 import fi.digitraffic.tis.vaco.queuehandler.model.ValidationInput;
 import org.postgresql.util.PGobject;
 import org.slf4j.Logger;
